% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_explorer_export.R
\name{export_to_decision_explorer}
\alias{export_to_decision_explorer}
\title{Exports a network as an xml-formatted Decision Explorer model}
\usage{
export_to_decision_explorer(network, filepath = NULL, scaling = 5)
}
\arguments{
\item{network}{The network to be exported}

\item{filepath}{The filepath to save the xml file to. If NULL, the xml document is returned without being saved}

\item{scaling}{Factor to scale the coordinates up by}
}
\value{
An xml document formatted to be imported into Decision Explorer
}
\description{
Exports a network as an xml-formatted Decision Explorer model
}
\examples{
export_from_decision_explorer_xml(example_network, file = "example_network.mdx", scaling = 5)
}
