% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumu_export.R
\name{export_to_kumu}
\alias{export_to_kumu}
\title{Exports a network as a json-formatted Kumu model}
\usage{
export_to_kumu(network, filepath = NULL, scaling = 2)
}
\arguments{
\item{network}{The network to be exported}

\item{filepath}{The filepath to save the json file to. If NULL, the json string is returned without being saved}

\item{scaling}{Factor to scale the coordinates up by}
}
\value{
A json string formatted to be imported into Kumu
}
\description{
Exports a network as a json-formatted Kumu model
}
\examples{
export_to_kumu(example_network, file = "example_network.json", scaling = 2)
}
