% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzer_export.R
\name{export_to_netzer}
\alias{export_to_netzer}
\title{Exports a network as csvs for use in NetZer}
\usage{
export_to_netzer(network, filepath = NULL)
}
\arguments{
\item{network}{The network to be exported}

\item{filepath}{The filepath to save the csvs file to. Two csvs are created, inserting "_nodes" and "_edges" ahead of the filetype. If NULL, the list of nodes and edges is returned without being saved}

\item{scaling}{Factor to scale the coordinates up by}
}
\value{
A list containing nodes and edges formatted for NetZer
}
\description{
Exports a network as csvs for use in NetZer
}
\examples{
export_to_netzer(alternet::example_network, file = "example_network.csv")
}
