% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stella_export.R
\name{export_to_stella}
\alias{export_to_stella}
\title{Exports a network as an xml-formatted Stella CLD}
\usage{
export_to_stella(network, filepath = NULL, translation = c(200, 0))
}
\arguments{
\item{network}{The network to be exported}

\item{filepath}{The filepath to save the xml file to. If NULL, the xml document is returned without being saved}

\item{translation}{Number to translate the coordinate up by}
}
\value{
An xml document formatted to be imported into Stellar
}
\description{
Exports a network as an xml-formatted Stella CLD
}
\examples{
export_to_stella(example_network, file = "example_network.stmx", translation = c(200,0))
}
