% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stella_import.R
\name{get_edge_info_stella}
\alias{get_edge_info_stella}
\title{Extracts edge information from Stella xml CLD map (as created by xml2::read_xml())}
\usage{
get_edge_info_stella(raw_xml)
}
\arguments{
\item{raw_xml}{xml document read from the model file}
}
\value{
A tibble containing information about edges
}
\description{
Extracts edge information from Stella xml CLD map (as created by xml2::read_xml())
}
\examples{
xml2::read_xml("example_network.stmx") \%>\% get_edge_info_stella()
}
