% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stella_import.R
\name{get_edge_node_sf}
\alias{get_edge_node_sf}
\title{Extracts edge information from Stella xml stock and flow map (as created by xml2::read_xml())}
\usage{
get_edge_node_sf(raw_xml, end)
}
\arguments{
\item{raw_xml}{xml document read from the model file}

\item{end}{whether the node sought is the "from" or "to" end}
}
\value{
A vector containing the from or to nodes
}
\description{
Extracts edge information from Stella xml stock and flow map (as created by xml2::read_xml())
}
\examples{
xml2::read_xml("example_network.stmx") \%>\% get_edge_node_sf("from")
}
