% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stella_import.R
\name{get_edge_polarity_sf}
\alias{get_edge_polarity_sf}
\title{Extracts edge information from Stella xml stock and flow map (as created by xml2::read_xml())}
\usage{
get_edge_polarity_sf(raw_xml)
}
\arguments{
\item{raw_xml}{xml document read from the model file}
}
\value{
A vector containing the polarities
}
\description{
Extracts edge information from Stella xml stock and flow map (as created by xml2::read_xml())
}
\examples{
xml2::read_xml("example_network.stmx") \%>\% get_edge_polarity_sf()
}
