% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stella_import.R
\name{get_node_info_stella}
\alias{get_node_info_stella}
\title{Extracts node information from Stella xml (as created by xml2::read_xml())}
\usage{
get_node_info_stella(raw_xml, model_type)
}
\arguments{
\item{raw_xml}{xml document read from the model file}

\item{model_type}{Specify "cld" to read from a CLD model with variables as nodes or "sf" to read from a stock and flow model with modules as nodes}
}
\value{
A tibble containing information about nodes
}
\description{
Extracts node information from Stella xml (as created by xml2::read_xml())
}
\examples{
xml2::read_xml("example_network.stmx") \%>\% get_node_info_stella()
}
