% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stella_import.R
\name{get_node_info_stella_cld}
\alias{get_node_info_stella_cld}
\title{Extracts node information from Stella xml CLD map (as created by xml2::read_xml())}
\usage{
get_node_info_stella_cld(raw_xml)
}
\arguments{
\item{raw_xml}{xml document read from the model file}
}
\value{
A tibble containing information about nodes
}
\description{
Extracts node information from Stella xml CLD map (as created by xml2::read_xml())
}
\examples{
xml2::read_xml("example_network.stmx") \%>\% get_node_info_stella_cld()
}
