% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_explorer_import.R
\name{get_node_style_info_de}
\alias{get_node_style_info_de}
\title{Extracts node style information from Decision Explorer xml (as created by xml2::read_xml())}
\usage{
get_node_style_info_de(raw_xml)
}
\arguments{
\item{raw_xml}{xml document read from the model file}
}
\value{
A tibble containing information about styles
}
\description{
Extracts node style information from Decision Explorer xml (as created by xml2::read_xml())
}
\examples{
xml2::read_xml("example_network.mdx") \%>\% get_node_style_info_de()
}
