% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_explorer_import.R
\name{import_from_decision_explorer_xml}
\alias{import_from_decision_explorer_xml}
\title{Imports a Decision Explorer xml-formatted model as a network, converting to a standardised R format}
\usage{
import_from_decision_explorer_xml(filepath, scaling = 5)
}
\arguments{
\item{filepath}{The filepath of the Decision Explorer xml file}

\item{scaling}{Factor to scale the coordinates down by}
}
\value{
A named list with a tibble containing information about nodes, and a tibble containing information about edges
}
\description{
Imports a Decision Explorer xml-formatted model as a network, converting to a standardised R format
}
\examples{
import_from_decision_explorer_xml("inst/extdata/example_network.mdx", scaling = 5)
}
