% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumu_import.R
\name{import_from_kumu_json}
\alias{import_from_kumu_json}
\title{Imports a Kumu json-formatted model as a network, converting to a standardised R format}
\usage{
import_from_kumu_json(filepath, scaling = 2)
}
\arguments{
\item{filepath}{The filepath of the Kumu json file}

\item{scaling}{Factor to scale the coordinates down by}
}
\value{
A named list with a tibble containing information about nodes, and a tibble containing information about edges
}
\description{
Imports a Kumu json-formatted model as a network, converting to a standardised R format
}
\examples{
import_from_kumu_json("inst/extdata/example_network.json", scaling = 2)
}
