% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsm_import.R
\name{import_from_prsm_xlsx}
\alias{import_from_prsm_xlsx}
\title{Imports a PRSM excel-formatted model as a network, converting to a standardised R format}
\usage{
import_from_prsm_xlsx(filepath, scaling = 2)
}
\arguments{
\item{filepath}{The filepath of the PRSM Excel file}

\item{scaling}{Factor to scale the coordinates down by}
}
\value{
A named list with a tibble containing information about nodes, and a tibble containing information about edges
}
\description{
Imports a PRSM excel-formatted model as a network, converting to a standardised R format
}
\examples{
import_from_prsm_xlsx("inst/extdata/example_network.xlsx", scaling = 2)
}
