% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stella_import.R
\name{get_node_info_stella}
\alias{get_node_info_stella}
\title{Extracts node information from Decision Explorer xml (as created by xml2::read_xml())}
\usage{
get_node_info_stella(raw_xml)
}
\arguments{
\item{raw_xml}{xml document read from the model file}
}
\value{
A tibble containing information about nodes
}
\description{
Extracts node information from Decision Explorer xml (as created by xml2::read_xml())
}
\examples{
xml2::read_xml("example_network.mdx") \%>\% get_node_info_de()
}
